requirements.txt
discord.py>=2.4
aiohttp>=3.9
pandas>=2.2
numpy>=1.26
scipy>=1.13
matplotlib>=3.9
anthropic>=0.40
python-dotenv>=1.0
aiosqlite>=0.20
tastytrade>=8.0
fastapi>=0.115
uvicorn>=0.34
pyyaml>=6.0
pyarrow>=14.0
hmmlearn>=0.3.0
torch>=2.0
transformers>=4.40
playwright>=1.40
playwright-stealth>=1.0
